# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bipupu",
    "description": "BIPUPU API 服务",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/messages/": {
      "post": {
        "tags": [
          "消息",
          "消息"
        ],
        "summary": "Send Message",
        "description": "发送消息\n\n支持：\n- 用户间传讯（receiver_id 为用户的 bipupu_id）\n- 向服务号发送消息（receiver_id 为服务号 ID，如 \"cosmic.fortune\"）",
        "operationId": "send_message_api_messages__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "消息",
          "消息"
        ],
        "summary": "Get Messages",
        "description": "获取消息列表\n\n参数：\n- direction: sent（发件箱）或 received（收件箱）\n- page: 页码\n- page_size: 每页数量",
        "operationId": "get_messages_api_messages__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "sent 或 received",
              "default": "received",
              "title": "Direction"
            },
            "description": "sent 或 received"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/favorites": {
      "get": {
        "tags": [
          "消息",
          "消息"
        ],
        "summary": "Get Favorites",
        "description": "获取收藏消息列表",
        "operationId": "get_favorites_api_messages_favorites_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/{message_id}/favorite": {
      "post": {
        "tags": [
          "消息",
          "消息"
        ],
        "summary": "Add Favorite",
        "description": "收藏消息",
        "operationId": "add_favorite_api_messages__message_id__favorite_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/FavoriteCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Favorite Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "消息",
          "消息"
        ],
        "summary": "Remove Favorite",
        "description": "取消收藏",
        "operationId": "remove_favorite_api_messages__message_id__favorite_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/{message_id}": {
      "delete": {
        "tags": [
          "消息",
          "消息"
        ],
        "summary": "Delete Message",
        "description": "删除消息（只能删除收到的消息）",
        "operationId": "delete_message_api_messages__message_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/messages": {
      "get": {
        "tags": [
          "管理后台"
        ],
        "summary": "Admin Messages",
        "description": "消息管理页面",
        "operationId": "admin_messages_admin_messages_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FavoriteCreate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "FavoriteCreate"
      },
      "FavoriteListResponse": {
        "properties": {
          "favorites": {
            "items": {
              "$ref": "#/components/schemas/FavoriteResponse"
            },
            "type": "array",
            "title": "Favorites"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "favorites",
          "total",
          "page",
          "page_size"
        ],
        "title": "FavoriteListResponse"
      },
      "FavoriteResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "message_id": {
            "type": "integer",
            "title": "Message Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "message_id",
          "user_id",
          "note",
          "created_at"
        ],
        "title": "FavoriteResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageCreate": {
        "properties": {
          "receiver_id": {
            "type": "string",
            "title": "Receiver Id",
            "description": "接收者的 bipupu_id"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "消息内容"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type",
            "description": "消息类型：NORMAL, VOICE, SYSTEM",
            "default": "NORMAL"
          },
          "pattern": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern",
            "description": "json扩展字段"
          }
        },
        "type": "object",
        "required": [
          "receiver_id",
          "content"
        ],
        "title": "MessageCreate",
        "description": "创建消息 - BIPU机消息发送格式"
      },
      "MessageListResponse": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "type": "array",
            "title": "Messages"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "messages",
          "total",
          "page",
          "page_size"
        ],
        "title": "MessageListResponse",
        "description": "消息列表响应 - BIPU机消息分页查询"
      },
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "sender_bipupu_id": {
            "type": "string",
            "title": "Sender Bipupu Id"
          },
          "receiver_bipupu_id": {
            "type": "string",
            "title": "Receiver Bipupu Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type"
          },
          "pattern": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sender_bipupu_id",
          "receiver_bipupu_id",
          "content",
          "message_type",
          "created_at"
        ],
        "title": "MessageResponse",
        "description": "消息响应 - BIPU机消息完整信息"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}